Implemented new/updated requirements.

+++ theme/single-chemical_element.php
@@ -3,6 +3,6 @@
 use Zakjakub\ChemieZijeTheme\Model\ChemicalElement;
 
 $context = Timber::context();
-$context['element'] = ChemicalElement::getElement(carbon_get_the_post_meta('proton_number'));
+$context['element'] = ChemicalElement::getElement((int)carbon_get_the_post_meta('proton_number'));
 $templates = ['single-customs/single-chemical_element.html.twig', 'post-types/page-layout.html.twig'];
 Timber::render($templates, $context);
